ReportActivity now uses Controller
