+++ Kuroba/app/src/main/java/com/github/adamantcheese/chan/ui/controller/settings/MainSettingsController.java
@@ -22,7 +22,7 @@ import com.github.adamantcheese.chan.BuildConfig;
 import com.github.adamantcheese.chan.R;
 import com.github.adamantcheese.chan.StartActivity;
 import com.github.adamantcheese.chan.core.presenter.SettingsPresenter;
-import com.github.adamantcheese.chan.report.ReportActivity;
+import com.github.adamantcheese.chan.report.ReportProblemController;
 import com.github.adamantcheese.chan.ui.controller.FiltersController;
 import com.github.adamantcheese.chan.ui.controller.LicensesController;
 import com.github.adamantcheese.chan.ui.controller.SitesSetupController;
@@ -30,6 +30,8 @@ import com.github.adamantcheese.chan.ui.settings.LinkSettingView;
 import com.github.adamantcheese.chan.ui.settings.SettingsGroup;
 
 import javax.inject.Inject;
+
+import kotlin.Unit;
 
 import static com.github.adamantcheese.chan.Chan.inject;
 import static com.github.adamantcheese.chan.utils.AndroidUtils.getApplicationLabel;
@@ -166,7 +168,16 @@ public class MainSettingsController
                         this,
                         "Report",
                         "Report a problem/crash",
-                        v -> ReportActivity.startActivity(context)
+                        v -> {
+                            ReportProblemController reportProblemController
+                                    = new ReportProblemController(context);
+
+                            reportProblemController.setOnFinishedCallback(() -> {
+                                reportProblemController.stopPresenting();
+                                return Unit.INSTANCE;
+                            });
+                            navigationController.presentController(reportProblemController);
+                        }
                 )
         );
